taking plain css out of tail wind css create file
We need to keep vanilla css out of the tailwind source file, If you want to create something specific you cannot do with tailwind inline, read the documentation on how to create a variable to use. The source file is used to render the stylesheets/style.css and during the purge tailwindcss for production I am not sure what would happen.